+++ supabase/seed.sql
@@ -380,7 +380,8 @@ insert into
 		"currency",
 		"welcome_amount",
 		"submission_cost",
-		"admins"
+		"admins",
+		"inactive"
 	)
 values
 	(
@@ -391,7 +392,8 @@ values
 		'c60c9fca-ad37-11f0-a9a1-57b72e1e85ac',
 		'20',
 		'10',
-		'{"d181d165-8b6a-4d79-ad28-a9aece21d813"}'
+		'{"d181d165-8b6a-4d79-ad28-a9aece21d813"}',
+		null
 	);
 
 insert into